+++ broccoli_server/contrib/one_off_job/__init__.py
@@ -1,1 +1,1 @@
-from .example import ExampleOneOffJob
+from .example import ExampleJob
