+++ broccoli_server/contrib/one_off_job/example.py
@@ -1,13 +1,13 @@
 import time
-from broccoli_server.interface.one_off_job import OneOffJob, OneOffJobContext
+from broccoli_server.interface.job import Job, JobContext
 
 
-class ExampleOneOffJob(OneOffJob):
+class ExampleJob(Job):
     def __init__(self, wait_seconds: int, print_str: str):
         self.wait_seconds = wait_seconds
         self.print_str = print_str
 
-    def work(self, context: OneOffJobContext):
+    def work(self, context: JobContext):
         time.sleep(self.wait_seconds)
         context.logger().info(f"{context.content_store().count({})} items in content store")
         context.logger().info(f"print_str={self.print_str}")
