+++ broccoli_server/interface/job/job.py
@@ -1,0 +1,8 @@
+from abc import ABCMeta, abstractmethod
+from .job_context import JobContext
+
+
+class Job(metaclass=ABCMeta):
+    @abstractmethod
+    def work(self, context: JobContext):
+        pass
