--- broccoli_server/interface/one_off_job/one_off_job.py
@@ -1,8 +1,0 @@
-from abc import ABCMeta, abstractmethod
-from .one_off_job_context import OneOffJobContext
-
-
-class OneOffJob(metaclass=ABCMeta):
-    @abstractmethod
-    def work(self, context: OneOffJobContext):
-        pass
