+++ broccoli_server/job/__init__.py
@@ -1,3 +1,3 @@
-from .one_off_job_executor import OneOffJobExecutor
+from .job_scheduler import JobScheduler
 from .job_run import JobRun
 from .job_runs_store import JobRunsStore
